requirements.txt
#Modulos Necesarios Para utilizar este bot:
numpy
discord
googlesearch-python